gsubprocess: Add UTF-8 variants of communicate()
Over many years of writing code interacting with subprocesses,
a pattern that comes up a lot is to run a child and get its
output as UTF-8, to put inside a JSON document or render
in a GtkTextBuffer, etc.

It's very important to validate at the boundaries, and not
say deep inside Pango.

We could do this a bit more efficiently if done in a streaming
fashion, but realistically this should be OK for now.